modify epoch params type to uint64

+++ ethclient/ethclient_map3.go
@@ -56,13 +56,13 @@ func (ec *Client) GetAllMap3RewardByDelegatorAddress(ctx context.Context, delega
 	return (*big.Int)(&result), err
 }
 
-func (ec *Client) GetActiveMap3NodeAtEpoch(ctx context.Context, epoch uint) ([]string, error) {
+func (ec *Client) GetActiveMap3NodeAtEpoch(ctx context.Context, epoch uint64) ([]string, error) {
 	var result []string
 	err := ec.c.CallContext(ctx, &result, "eth_getActiveMap3NodeAtEpoch", epoch)
 	return result, err
 }
 
-func (ec *Client) GetTerminatedMap3NodeAtEpoch(ctx context.Context, epoch uint) ([]string, error) {
+func (ec *Client) GetTerminatedMap3NodeAtEpoch(ctx context.Context, epoch uint64) ([]string, error) {
 	var result []string
 	err := ec.c.CallContext(ctx, &result, "eth_getTerminatedMap3NodeAtEpoch", epoch)
 	return result, err
